docs(cli-forge): convert oneOf to multi-file example with prose
Replaces single-file example with a multi-file example directory
including content.md prose, a color-output demo (boolean+string),
and a verbosity demo (number+string priority ordering).
